requirements.txt
fastapi
pymysql
python-dotenv 
uvicorn